requirements.txt
Flask>=2.0
numpy>=1.20
Pillow>=8.0
scipy>=1.7
matplotlib>=3.4
uuid
waitress